+++ segmentation/__init__.py
@@ -1,2 +1,3 @@
+from .utils import *
 from .data import load_data, generate_dataset
 from .model import load_model
